make useRoutes a function rather than hook
