+++ package.json
@@ -22,7 +22,6 @@
     "@reduxjs/toolkit": "^1.6.1",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^12.0.0",
-    "@testing-library/react-hooks": "^7.0.1",
     "@typechain/ethers-v5": "^7.0.0",
     "@types/d3": "^6.7.1",
     "@types/jest": "^25.2.1",
